splitting beamline motion from detector
