+++ tomo2bm/scan.py
@@ -21,6 +21,7 @@ import numpy as np
 from tomo2bm import aps2bm
 from tomo2bm import dm
 from tomo2bm import log
+from tomo2bm import flir
 
 global_PVs = {}
 
@@ -41,11 +42,11 @@ def fly_sleep(params):
             # calling global_PVs['Cam1_AcquireTime'] to replace the default 'ExposureTime' with the one set in the camera
             params.exposure_time = global_PVs['Cam1_AcquireTime'].get()
             # calling calc_blur_pixel() to replace the default 'SlewSpeed' 
-            blur_pixel, rot_speed, scan_time = aps2bm.calc_blur_pixel(global_PVs, params)
+            blur_pixel, rot_speed, scan_time = calc_blur_pixel(global_PVs, params)
             params.slew_speed = rot_speed
 
             # init camera
-            aps2bm.pgInit(global_PVs, params)
+            flir.pgInit(global_PVs, params)
 
             # set sample file name
             fname = str('{:03}'.format(global_PVs['HDF1_FileNumber'].get())) + '_' + global_PVs['Sample_Name'].get(as_string=True)
@@ -76,7 +77,7 @@ def dummy_tomo_fly_scan(global_PVs, params, fname):
     log.info('  *** start_scan')
 
     def cleanup(signal, frame):
-        aps2bm.stop_scan(global_PVs, params)
+        stop_scan(global_PVs, params)
         sys.exit(0)
     signal.signal(signal.SIGINT, cleanup)
 
@@ -86,12 +87,12 @@ def tomo_fly_scan(global_PVs, params, fname):
     log.info('  *** start_scan')
 
     def cleanup(signal, frame):
-        aps2bm.stop_scan(global_PVs, params)
+        stop_scan(global_PVs, params)
         sys.exit(0)
     signal.signal(signal.SIGINT, cleanup)
 
     # if params.has_key('StopTheScan'):
-    #     aps2bm.stop_scan(global_PVs, params)
+    #     stop_scan(global_PVs, params)
     #     return
 
     # moved to outer loop in main()
@@ -102,24 +103,147 @@ def tomo_fly_scan(global_PVs, params, fname):
     # fname = global_PVs['HDF1_FileName'].get(as_string=True)
     log.info('  *** File name prefix: %s' % fname)
 
-    aps2bm.pgSet(global_PVs, params, fname) 
+    flir.pgSet(global_PVs, params, fname) 
 
     aps2bm.open_shutters(global_PVs, params)
 
+    move_sample_in(global_PVs, params)
+
     # # run fly scan
-    theta = aps2bm.pgAcquisition(global_PVs, params)
+    theta = flir.pgAcquisition(global_PVs, params)
 
     theta_end =  global_PVs['Motor_SampleRot_RBV'].get()
     if (0 < theta_end < 180.0):
         # print('\x1b[2;30;41m' + '  *** Rotary Stage ERROR. Theta stopped at: ***' + theta_end + '\x1b[0m')
         log.error('  *** Rotary Stage ERROR. Theta stopped at: %s ***' % str(theta_end))
 
-    aps2bm.pgAcquireFlat(global_PVs, params)
+    move_sample_out(global_PVs, params)
+    flir.pgAcquireFlat(global_PVs, params)
+    move_sample_in(global_PVs, params)
+
     aps2bm.close_shutters(global_PVs, params)
     time.sleep(2)
 
-    aps2bm.pgAcquireDark(global_PVs, params)
+    flir.pgAcquireDark(global_PVs, params)
 
-    aps2bm.checkclose_hdf(global_PVs, params)
+    flir.checkclose_hdf(global_PVs, params)
 
-    aps2bm.add_theta(global_PVs, params, theta)
+    flir.add_theta(global_PVs, params, theta)
+
+
+########################################################################
+def calc_blur_pixel(global_PVs, params):
+    """
+    Calculate the blur error (pixel units) due to a rotary stage fly scan motion durng the exposure.
+    
+    Parameters
+    ----------
+    params.exposure_time: float
+        Detector exposure time
+    params.ccd_readout : float
+        Detector read out time
+    variableDict[''roiSizeX''] : int
+        Detector X size
+    params.sample_rotation_end : float
+        Tomographic scan angle end
+    params.sample_rotation_start : float
+        Tomographic scan angle start
+    variableDict[''Projections'] : int
+        Numember of projections
+
+    Returns
+    -------
+    float
+        Blur error in pixel. For good quality reconstruction this should be < 0.2 pixel.
+    """
+
+    angular_range =  params.sample_rotation_end -  params.sample_rotation_start
+    angular_step = angular_range/params.num_projections
+    scan_time = params.num_projections * (params.exposure_time + params.ccd_readout)
+    rot_speed = angular_range / scan_time
+    frame_rate = params.num_projections / scan_time
+    blur_delta = params.exposure_time * rot_speed
+ 
+   
+    mid_detector = global_PVs['Cam1_MaxSizeX_RBV'].get() / 2.0
+    blur_pixel = mid_detector * (1 - np.cos(blur_delta * np.pi /180.))
+
+    log.info(' ')
+    log.info('  *** Calc blur pixel')
+    log.info("  *** *** Total # of proj: %s " % params.num_projections)
+    log.info("  *** *** Exposure Time: %s s" % params.exposure_time)
+    log.info("  *** *** Readout Time: %s s" % params.ccd_readout)
+    log.info("  *** *** Angular Range: %s degrees" % angular_range)
+    log.info("  *** *** Camera X size: %s " % global_PVs['Cam1_SizeX'].get())
+    log.info(' ')
+    log.info("  *** *** *** *** Angular Step: %f degrees" % angular_step)   
+    log.info("  *** *** *** *** Scan Time: %f s" % scan_time) 
+    log.info("  *** *** *** *** Rot Speed: %f degrees/s" % rot_speed)
+    log.info("  *** *** *** *** Frame Rate: %f fps" % frame_rate)
+    log.info("  *** *** *** *** Max Blur: %f pixels" % blur_pixel)
+    log.info('  *** Calc blur pixel: Done!')
+    
+    return blur_pixel, rot_speed, scan_time
+
+
+def move_sample_out(global_PVs, params):
+    log.info('      *** Sample out')
+    if not (params.sample_move_freeze):
+        if (params.sample_in_out_vertical):
+            log.info('      *** *** Move Sample Y out at: %f' % params.sample_out_position)
+            global_PVs['Motor_SampleY'].put(str(params.sample_out_position), wait=True, timeout=1000.0)                
+            if wait_pv(global_PVs['Motor_SampleY'], float(params.sample_out_position), 60) == False:
+                log.error('Motor_SampleY did not move in properly')
+                log.error(global_PVs['Motor_SampleY'].get())
+        else:
+            if (params.use_furnace):
+                log.info('      *** *** Move Furnace Y out at: %f' % params.furnace_out_position)
+                global_PVs['Motor_FurnaceY'].put(str(params.furnace_out_position), wait=True, timeout=1000.0)
+                if wait_pv(global_PVs['Motor_FurnaceY'], float(params.furnace_out_position), 60) == False:
+                    log.error('Motor_FurnaceY did not move in properly')
+                    log.error(global_PVs['Motor_FurnaceY'].get())
+            log.info('      *** *** Move Sample X out at: %f' % params.sample_out_position)
+            global_PVs['Motor_SampleX'].put(str(params.sample_out_position), wait=True, timeout=1000.0)
+            if wait_pv(global_PVs['Motor_SampleX'], float(params.sample_out_position), 60) == False:
+                log.error('Motor_SampleX did not move in properly')
+                log.error(global_PVs['Motor_SampleX'].get())
+    else:
+        log.info('      *** *** Sample Stack is Frozen')
+
+
+def move_sample_in(global_PVs, params):
+    log.info('      *** Sample in')
+    if not (params.sample_move_freeze):
+        if (params.sample_in_out_vertical):
+            log.info('      *** *** Move Sample Y in at: %f' % params.sample_in_position)
+            global_PVs['Motor_SampleY'].put(str(params.sample_in_position), wait=True, timeout=1000.0)                
+            if wait_pv(global_PVs['Motor_SampleY'], float(params.sample_in_position), 60) == False:
+                log.error('Motor_SampleY did not move in properly')
+                log.error(global_PVs['Motor_SampleY'].get())
+        else:
+            log.info('      *** *** Move Sample X in at: %f' % params.sample_in_position)
+            global_PVs['Motor_SampleX'].put(str(params.sample_in_position), wait=True, timeout=1000.0)
+            if wait_pv(global_PVs['Motor_SampleX'], float(params.sample_in_position), 60) == False:
+                log.error('Motor_SampleX did not move in properly')
+                log.error(global_PVs['Motor_SampleX'].get())
+            if (params.use_furnace):
+                log.info('      *** *** Move Furnace Y in at: %f' % params.furnace_in_position)
+                global_PVs['Motor_FurnaceY'].put(str(params.furnace_in_position), wait=True, timeout=1000.0)
+                if wait_pv(global_PVs['Motor_FurnaceY'], float(params.furnace_in_position), 60) == False:
+                    log.error('Motor_FurnaceY did not move in properly')
+                    log.error(global_PVs['Motor_FurnaceY'].get())
+    else:
+        log.info('      *** *** Sample Stack is Frozen')
+
+def stop_scan(global_PVs, params):
+        log.info(' ')
+        log.error('  *** Stopping the scan: PLEASE WAIT')
+        global_PVs['Motor_SampleRot_Stop'].put(1)
+        global_PVs['HDF1_Capture'].put(0)
+        wait_pv(global_PVs['HDF1_Capture'], 0)
+        pgInit(global_PVs, params)
+        log.error('  *** Stopping scan: Done!')
+        ##pgInit(global_PVs, params)
+
+########################################################################
+
